requirements.txt
asgiref==3.2.3
autopep8==1.4.4
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
pep8==1.7.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
